+++ packages/create-pantheon-decoupled-kit/__tests__/testGenerators/generators/index.ts
@@ -1,3 +1,0 @@
-import { testAdd } from './test.generator';
-import { testAppend } from './test-append.generator';
-export const decoupledKitTestGenerators = [testAdd, testAppend];