+++ packages/create-pantheon-decoupled-kit/src/generators/gatsby-wp-acf-addon.generator.ts
@@ -1,7 +1,7 @@
-import type { CustomActionConfig } from 'node-plop';
-import type { DecoupledKitGenerator } from '../types';
+import { addWithDiff, runLint } from '../actions';
+import type { DecoupledKitGenerator, DefaultAnswers } from '../types';
 
-export const gatsbyWpAcfAddon: DecoupledKitGenerator = {
+export const gatsbyWpAcfAddon: DecoupledKitGenerator<DefaultAnswers> = {
 	name: 'gatsby-wp-acf-addon',
 	description:
 		'Example implementation of the WordPress Advanced Custom Fields plugin for the gatsby-wordpress starter',
@@ -9,22 +9,9 @@ export const gatsbyWpAcfAddon: DecoupledKitGenerator = {
 		{
 			name: 'outDir',
 			message: 'Where should the output go?',
-			default: `${process.cwd()}/acf-demo`,
+			default: `${process.cwd()}/gatsby-wp-acf-addon`,
 		},
 	],
-	actions: (data) => {
-		const addWithDiff: CustomActionConfig<'addWithDiff'> = {
-			type: 'addWithDiff',
-			templates: './templates/gatsby-wp-acf-addon',
-			path: '{{outDir}}',
-			force: data?.force ? Boolean(data.force) : false,
-		};
-		const runESLint: CustomActionConfig<'runLint'> = {
-			type: 'runLint',
-		};
-
-		const actions = [addWithDiff, runESLint];
-
-		return actions;
-	},
+	templates: ['gatsby-wp-acf-addon'],
+	actions: [addWithDiff, runLint],
 };